more accurate audio timestamps. Patch by Mark Spieth. Improve A/V Sync. Closes #8500
git-svn-id: http://svn.mythtv.org/svn/trunk@26255 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmyth/audiooutput.h
@@ -4,6 +4,7 @@
 #include <QString>
 #include <QVector>
 
+#include "compat.h"
 #include "audiosettings.h"
 #include "audiooutputsettings.h"
 #include "mythcorecontext.h"
@@ -70,9 +71,9 @@ class MPUBLIC AudioOutput : public VolumeBase, public OutputListeners
 
     virtual void Reset(void) = 0;
 
-    virtual bool AddFrames(void *buffer, int samples, long long timecode) = 0;
+    virtual bool AddFrames(void *buffer, int samples, int64_t timecode) = 0;
 
-    virtual void SetTimecode(long long timecode) = 0;
+    virtual void SetTimecode(int64_t timecode) = 0;
     virtual bool IsPaused(void) const = 0;
     virtual void Pause(bool paused) = 0;
     virtual void PauseUntilBuffered(void) = 0;
@@ -80,10 +81,10 @@ class MPUBLIC AudioOutput : public VolumeBase, public OutputListeners
     // Wait for all data to finish playing
     virtual void Drain(void) = 0;
 
-    virtual int GetAudiotime(void) = 0;
+    virtual int64_t GetAudiotime(void) = 0;
 
     /// report amount of audio buffered in milliseconds.
-    virtual int GetAudioBufferedTime(void) { return 0; }
+    virtual int64_t GetAudioBufferedTime(void) { return 0; }
 
     virtual void SetSourceBitrate(int ) { }
